Added conftest.py to centralize pytest fixtures for unit and integration tests
