+++ tests/unit/test_alias.py
@@ -1,18 +1,6 @@
-from pathlib import Path
-
 import pytest
 
 from cromshell.alias import command as alias_command
-
-
-@pytest.fixture
-def mock_data_path():
-    return Path(__file__).parent.joinpath("mock_data/")
-
-
-@pytest.fixture
-def submission_file(mock_data_path):
-    return mock_data_path.joinpath("all.workflow.database.tsv")
 
 
 class TestAlias:
@@ -44,11 +32,11 @@ class TestAlias:
         self,
         alias_name: str,
         alias_existence: bool,
-        submission_file,
+        mock_workflow_database_tsv,
     ) -> None:
         assert (
             alias_command.alias_exists(
-                alias_name=alias_name, submission_file=str(submission_file)
+                alias_name=alias_name, submission_file=str(mock_workflow_database_tsv)
             )
             is alias_existence
         )
@@ -62,10 +50,10 @@ class TestAlias:
     )
     #  'caplog' is pytest fixture
     def test_check_workflow_has_alias(
-        self, workflow_id, alias_existence, submission_file, caplog
+        self, workflow_id, alias_existence, mock_workflow_database_tsv, caplog
     ) -> None:
         alias_command.check_workflow_has_alias(
-            workflow_id=workflow_id, submission_file=submission_file
+            workflow_id=workflow_id, submission_file=mock_workflow_database_tsv
         )
         for record in caplog.records:
             assert record.levelname == "WARNING"
